// webapp/src/api/model/eAction.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type EAction = typeof EAction[keyof typeof EAction];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EAction = {
  ACTION: 'ACTION',
  BONUS_ACTION: 'BONUS_ACTION',
  MOVEMENT_ACTION: 'MOVEMENT_ACTION',
  REACTION: 'REACTION',
} as const;