// webapp/src/api/model/spell.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { EAction } from './eAction';
import type { SpellCast } from './spellCast';
import type { ECastingType } from './eCastingType';
import type { SpellCostItem } from './spellCostItem';

export interface Spell {
  action: EAction;
  additionalCost: string;
  cast: SpellCast;
  castTime?: number;
  castingTypes: ECastingType[];
  cooldown?: number;
  cost: SpellCostItem[];
  countermeasures: string;
  effect: string;
  id?: string;
  name: string;
  tags: string[];
  tier: number;
}